move gene image width constant to EV specific
