+++ src/ensembl/src/content/app/entity-viewer/gene-view/components/gene-overview-image/GeneOverviewImage.tsx
@@ -28,14 +28,14 @@ import { TicksAndScale } from 'src/content/app/entity-viewer/gene-view/component
 
 import styles from './GeneOverviewImage.scss';
 import { pluralise } from 'src/shared/helpers/formatters/pluralisationFormatter';
-import settings from 'src/styles/_settings.scss';
+import settings from 'src/content/app/entity-viewer/gene-view/styles/_constants.scss';
 
 type GeneOverviewImageProps = {
   gene: Gene;
   onTicksCalculated: (payload: TicksAndScale) => void;
 };
 
-const image_width = Number(settings.image_width);
+const gene_image_width = Number(settings.gene_image_width);
 
 const GeneOverviewImage = (props: GeneOverviewImageProps) => {
   const { start: geneStart, end: geneEnd } = getFeatureCoordinates(props.gene); // FIXME: use gene length further on
@@ -51,7 +51,7 @@ const GeneOverviewImage = (props: GeneOverviewImageProps) => {
       <div className={styles.ruler}>
         <BasePairsRuler
           length={length}
-          width={image_width}
+          width={gene_image_width}
           onTicksCalculated={props.onTicksCalculated}
           standalone={true}
         />
@@ -67,7 +67,7 @@ export const GeneImage = (props: GeneOverviewImageProps) => {
   // (it will help with drawing genes of circular chromosomes)
   const scale = scaleLinear()
     .domain([geneStart, geneEnd])
-    .range([0, image_width]);
+    .range([0, gene_image_width]);
 
   const renderedTranscripts = props.gene.transcripts.map(
     (transcript, index) => {
@@ -94,7 +94,7 @@ export const GeneImage = (props: GeneOverviewImageProps) => {
   );
 
   return (
-    <svg className={styles.containerSVG} width={image_width}>
+    <svg className={styles.containerSVG} width={gene_image_width}>
       {renderedTranscripts}
     </svg>
   );
