+++ src/ensembl/src/styles/_settings.scss
@@ -16,9 +16,6 @@ $global-radius: 0;
 $global-padding-left: 100px;
 $global-button-cursor: auto;
 
-// Image configurations
-$image_width: 695;
-
 // Colour Palette
 // ---------
 $soft-black: #374148;
@@ -59,7 +56,6 @@ $global-box-shadow: $medium-light-grey;
   green: $green;
   white: $white;
   shadow-color: $shadow-color;
-  image_width: $image_width;
 }
 /* stylelint-enable */
